# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Membership Management API Documentation",
        "description": "API documentation using Swagger",
        "version": "1.0"
    },
    "paths": {
...
        "/api/fields": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Lấy danh sách lĩnh vực",
                "operationId": "ed4e3b5962d776e2c809da06ef98bcb9",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm kiếm theo tên lĩnh vực",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Bao gồm thông tin ngành nghề (giá trị: industry)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "industry"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số lượng lĩnh vực trên mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách lĩnh vực",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Field"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fields"
                ],
                "summary": "Thêm lĩnh vực",
                "operationId": "5d9dc653b47441716ba2895eb04097e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFieldRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lĩnh vực đã được tạo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fields/{id}": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Lấy thông tin lĩnh vực",
                "operationId": "448c44f7fae786393ad58cf62a57fe15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết lĩnh vực",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Fields"
                ],
                "summary": "Cập nhật lĩnh vực",
                "operationId": "af9b47e04bce042a87e12812024e1d4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFieldRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật lĩnh vực thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fields"
                ],
                "summary": "Xóa lĩnh vực",
                "operationId": "fcd95fafaad253751c4c1b5a9ffa8401",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa lĩnh vực thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreFieldRequest": {
                "required": [
                    "name",
                    "code",
                    "industry_id"
                ],
                "properties": {
                    "name": {
                        "description": "Tên của lĩnh vực",
                        "type": "string",
                        "example": "Lập trình phần mềm"
                    },
                    "code": {
                        "description": "Mã lĩnh vực, phải là duy nhất",
                        "type": "string",
                        "example": "IT-SW"
                    },
                    "description": {
                        "description": "Mô tả lĩnh vực",
                        "type": "string",
                        "example": "Lĩnh vực phát triển phần mềm"
                    },
                    "industry_id": {
                        "description": "ID của ngành nghề liên quan",
                        "type": "integer",
                        "example": 1
                    },
                    "sub_groups": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "description": "Tên nhóm con",
                                    "type": "string",
                                    "example": "Frontend Development"
                                },
                                "description": {
                                    "description": "Mô tả nhóm con",
                                    "type": "string",
                                    "example": "Phát triển giao diện người dùng"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "Field": {
                "title": "Field",
                "description": "Thông tin lĩnh vực",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "IT-DEV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Phát triển phần mềm"
                    },
                    "description": {
                        "type": "string",
                        "example": "Lĩnh vực chuyên về phát triển phần mềm, ứng dụng và hệ thống"
                    },
                    "industry_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "industry": {
                        "$ref": "#/components/schemas/Industry"
                    }
                },
                "type": "object"
            },
            "Industry": {
                "title": "Industry",
                "description": "Thông tin ngành nghề",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "industry_code": {
                        "type": "string",
                        "example": "IT"
                    },
                    "industry_name": {
                        "type": "string",
                        "example": "Công nghệ thông tin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ngành liên quan đến phần mềm, phần cứng và công nghệ"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}